add convergence check and fix ladybug projection

+++ include/CeresCameraCostFunction.hpp
@@ -13,7 +13,7 @@ struct CameraProjectionFunctor {
 
   bool operator()(const double* P, double* pixel) const {
     Eigen::Vector3d P_CAMERA_eig{P[0], P[1], P[2]};
-    std::optional<Eigen::Vector2d> pixel_projected =
+    std::optional<Eigen::Vector2i> pixel_projected =
         camera_model_->ProjectPoint(P_CAMERA_eig);
     if (!pixel_projected.has_value()) { return false; }
     pixel[0] = pixel_projected.value()[0];
